+++ tests/parser.cc
@@ -417,7 +417,7 @@ TEST_CASE("correct use of Mata::Parser::parse_mf_section()")
 		REQUIRE(ref->size() == 0);
 
 		// check what remains
-		std::string remains = stream.str().substr(stream.tellg());
+		std::string remains = stream.str().substr(static_cast<unsigned long>(stream.tellg()));
 		REQUIRE("@Type2\n%key2\n" == remains);
 	}
 } // parse_mf_section correct }}}
